refactor: adopt proxychannel for calling INativeMcpDiscoveryHelperService
