+++ src/vs/platform/mcp/node/nativeMcpDiscoveryHelperChannel.ts
@@ -3,17 +3,17 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { homedir } from 'os';
 import { Event } from '../../../base/common/event.js';
-import { URI } from '../../../base/common/uri.js';
 import { IURITransformer, transformOutgoingURIs } from '../../../base/common/uriIpc.js';
 import { IServerChannel } from '../../../base/parts/ipc/common/ipc.js';
-import { INativeMcpDiscoveryData } from '../common/nativeMcpDiscoveryHelper.js';
-import { platform } from '../../../base/common/platform.js';
+import { INativeMcpDiscoveryHelperService } from '../common/nativeMcpDiscoveryHelper.js';
 
 export class NativeMcpDiscoveryHelperChannel implements IServerChannel {
 
-	constructor(private getUriTransformer: undefined | ((requestContext: any) => IURITransformer)) { }
+	constructor(
+		private getUriTransformer: undefined | ((requestContext: any) => IURITransformer),
+		@INativeMcpDiscoveryHelperService private nativeMcpDiscoveryHelperService: INativeMcpDiscoveryHelperService
+	) { }
 
 	listen(context: any, event: string): Event<any> {
 		throw new Error('Invalid listen');
@@ -23,25 +23,11 @@ export class NativeMcpDiscoveryHelperChannel implements IServerChannel {
 		const uriTransformer = this.getUriTransformer?.(context);
 		switch (command) {
 			case 'load': {
-				const result: INativeMcpDiscoveryData = {
-					platform,
-					homedir: URI.file(homedir()),
-					winAppData: this.uriFromEnvVariable('APPDATA'),
-					xdgHome: this.uriFromEnvVariable('XDG_CONFIG_HOME'),
-				};
-
+				const result = await this.nativeMcpDiscoveryHelperService.load();
 				return uriTransformer ? transformOutgoingURIs(result, uriTransformer) : result;
 			}
 		}
 		throw new Error('Invalid call');
-	}
-
-	private uriFromEnvVariable(varName: string) {
-		const envVar = process.env[varName];
-		if (!envVar) {
-			return undefined;
-		}
-		return URI.file(envVar);
 	}
 }
