+++ src/vs/workbench/contrib/mcp/electron-sandbox/nativeMpcDiscovery.ts
@@ -3,13 +3,13 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { ProxyChannel } from '../../../../base/parts/ipc/common/ipc.js';
 import { IFileService } from '../../../../platform/files/common/files.js';
 import { IInstantiationService } from '../../../../platform/instantiation/common/instantiation.js';
 import { IMainProcessService } from '../../../../platform/ipc/common/mainProcessService.js';
 import { ILabelService } from '../../../../platform/label/common/label.js';
 import { ILogService } from '../../../../platform/log/common/log.js';
-import { INativeMcpDiscoveryData, NativeMcpDiscoveryHelperChannelName } from '../../../../platform/mcp/common/nativeMcpDiscoveryHelper.js';
-import { Dto } from '../../../services/extensions/common/proxyIdentifier.js';
+import { INativeMcpDiscoveryHelperService, NativeMcpDiscoveryHelperChannelName } from '../../../../platform/mcp/common/nativeMcpDiscoveryHelper.js';
 import { FilesystemMpcDiscovery } from '../common/discovery/nativeMcpDiscoveryAbstract.js';
 import { IMcpRegistry } from '../common/mcpRegistryTypes.js';
 
@@ -26,14 +26,15 @@ export class NativeMcpDiscovery extends FilesystemMpcDiscovery {
 	}
 
 	public override start(): void {
-		this.mainProcess.getChannel(NativeMcpDiscoveryHelperChannelName)
-			.call<Dto<INativeMcpDiscoveryData>>('load', undefined)
-			.then(
-				data => this.setDetails(data),
-				err => {
-					this.logService.warn('Error getting main process MCP environment', err);
-					this.setDetails(undefined);
-				}
-			);
+		const service = ProxyChannel.toService<INativeMcpDiscoveryHelperService>(
+			this.mainProcess.getChannel(NativeMcpDiscoveryHelperChannelName));
+
+		service.load().then(
+			data => this.setDetails(data),
+			err => {
+				this.logService.warn('Error getting main process MCP environment', err);
+				this.setDetails(undefined);
+			}
+		);
 	}
 }
